merapikan struktur foder dan rename browser router
