+++ src/container/Pages/BlogPost/BlogPost.jsx
@@ -1,5 +1,5 @@
 import React, { Component, Fragment } from 'react';
-import Post from '../../component/Post/Post';
+import Post from '../../../component/Post/Post';
 import './BlogPost.css';
 import axios from 'axios';
