+++ src/container/Pages/Product/Product.jsx
@@ -1,5 +1,5 @@
 import React, { Component, Fragment } from 'react';
-import CardProduct from '../CardProduct/CardProduct';
+import CardProduct from '../../CardProduct/CardProduct';
 import './Product.css';
 
 class Product extends Component {
